feat: rename existing menu components to use new conventions from design
